feature/#3 - 검색화면 UI 구현
